+++ docs/components/FormContainer.js
@@ -14,8 +14,16 @@ const onSubmit = values => {
   return new Promise(resolve => setTimeout(() => resolve(), 1000));
 };
 
+const initialValues = {
+  cents: 523,
+  email: "kevin@culturehq.com",
+  number: 42,
+  password: "password",
+  string: "@culturehq/components"
+};
+
 const FormContainer = () => (
-  <Form onSubmit={onSubmit}>
+  <Form onSubmit={onSubmit} initialValues={initialValues}>
     <CentsField label="Cents" name="cents" required />
     <EmailField label="Email" name="email" required />
     <NumberField label="Number" name="number" required />