+++ src/components/Spinner.js
@@ -1,8 +1,12 @@
 import React from "react";
 import classnames from "classnames";
 
-const Spinner = ({ className }) => (
-  <div className={classnames("chq-spn", className)}><div /></div>
+const Spinner = ({ className, placeholder }) => (
+  <div
+    className={classnames("chq-spn", className, { "chq-spn-ph": placeholder })}
+  >
+    <div />
+  </div>
 );
 
 export default Spinner;